requirements.txt
altair==5.3.0
asgiref==3.8.1
attrs==23.2.0
blinker==1.8.2
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
dj-database-url==2.2.0
Django==5.0.6
django-ratelimit==4.1.0
fonttools==4.53.0
gitdb==4.0.11
GitPython==3.1.43
graphviz==0.20.3
gunicorn==22.0.0
h11==0.14.0
h5py==3.11.0
idna==3.7
imbalanced-learn==0.13.0
isodate==0.6.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.2
protobuf==5.27.2
psycopg2-binary==2.9.9
pyarrow==16.1.0
pycparser==2.22
pydeck==0.9.1
pydot==2.0.0
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
rapidminer==10.1.0
referencing==0.35.1
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.1
scikit-learn==1.6.1
scipy==1.14.0
seaborn==0.13.2
six==1.16.0
sklearn-compat==0.1.3
smmap==5.0.1
sqlparse==0.5.0
streamlit==1.36.0
tenacity==8.4.2
threadpoolctl==3.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.1
watchdog==4.0.1
whitenoise==6.7.0
zeep==4.2.1
